backend/requirements.txt
fastapi
uvicorn
polars
pandas
numpy
nltk
scikit-learn
python-multipart
python-dotenv
google-generativeai
python-dotenv
tqdm